requirements.txt
fastapi
uvicorn
python-dotenv


chromadb
unstructured


langchain
langchain-community
langchain-cohere
langchain_pinecone